feat(moment): reintroduce Moment.js as peer
It turns out that the convert function which among other things also
converts Moment dates is also available in data set. Yesterday I checked
all the projects except vis-data which was a really stupid mistake.
However I still don't want Moment bundled here since the other projects
can live without it and Timeline would end up with two copies of Moment.
This makes it a peer dependency to avoid that. It will be installed only
once in Timeline and optional in the other projects.