fix(build): create IE11 compatible ESM
Babel is "smart" and doesn't add IE11 polyfills to ESM since IE11 can't
load ESM anyway but then we go and bundle ESM into UMD. We can't add the
polyfills later because that would mean polyfilling polyfills so they
simply have to be there to begin with.

This commit simply forces Babel to ignore that it's polyfilling ESM and
act as if it was UMD which means include IE11 polyfills.