feat(layered-storage): remove events
BREAKING CHANGE: The on methods were removed from Layered Storage.

This was unbearably slow. Most of the time was actually spent handling
events. Making it faster without reducing it's functionality to the same
level as Emitter seems impossible to me. Since we already use the
Emitter there's no point in having such thing.