chore(package): update dependencies (#78)
After this update we are no longer going to require any of the vunerable
versions of Handlebars:

Versions of handlebars prior to 4.3.0 are vulnerable to Prototype
Pollution leading to Remote Code Execution. Templates may alter an
Object's proto and defineGetter properties, which may allow an attacker
to execute arbitrary code through crafted payloads.

PS: This was used only in tests and docs but not in production code
hence no need for a release.