style: replace deprecated JSDoc with plugins
The built in JSDoc in ESLint is deprecated for some time. This replaces
it with TSDoc plugin to validate syntax and JSDoc plugin for general
linting. This is much more powerful and configurable solution. Atop of
that the JSDoc plugin seems to be TypeScript aware to great extent which
is also a plus.

Many changes had to be made to the documentation because of this.