feat(random): add an RNG based on Alea (#82)
* feat(random): add good, seedable RNG

We often use some seedable RNGs based on sinus. There are several
problems with these:
- The values are generally highly correlated.
- Implementation differs and therefore fails pinning tests.
- **Vast majority** of the returned values is **very** close to 1.
- Values close to 0 are as good as imposibble.

This provides a solid alternatives that will hopefully be much better
and consistent in different environments (allowing for pinning tests).

* test(deep-merge): fix failing test

Creating a snapshot somehow made this fail. I suspect some kind of
global scope pollution. Sadly unless we want to switch to AVA or
something this seems to be the most elaborate and best maintained
snapshotting library.

Truth is though that the test should have always looked like this.