feat(uuid)!: don't reexport UUID
BREAKING CHANGE: There's no randomUUID function anymore.

It was simply renamed and reexported. I don't really see the advantage
in this. Timeline and Network will have to use the uuid package
directly:

```javascript
import { default as randomUUID } from "uuid/v4";
```

instead of

```javascript
import { randomUUID } from "vis-util";
```

We export this from our builds but AFAIK it's all undocumented. Do we
consider this to be a part of the public API? If so this should probably
be posponed and deprecation warning added in the mean time to give
people more time to adjust.