feat(build)!: use common build process (#144)
BREAKING CHANGE: The old bundles are no longer present in the package. It has to be replaced by import * as util from 'vis-util/esnext';. I'll do that in all of our projects so don't bother updating to this, it most likely won't work anyway.

The build process from visjs/vis-dev-utils#10 is finally being deployed tada.

I also increased the timeout of some tests because they kept failing solely due to not having enough of time to finish.