feat(deep-object-assign): add pure version
I thought we mutate the items during updates in Data Sets but apparently
we don't, we clone them (I really should do some interoperability
testing before I open PRs). This is basically `deepObjectAssign({},
base, update)` but with proper typesafety and more clearly communicating
the intention of not modifying the arguments.