refactor(shared): prepare shared folder and hammer module for extraction (#1255)
* refactor: rework all imports and exports to ESM

* refactor(color-picker): inline onTouch from hammerUtil

Similarly as with the Activator, it was used only for one very simple
thing so there isn't really a need to drag this over to Vis Util.

* refactor(hammer): move it into shared

Everything from shared will be moved to Vis Util so that it can be
shared among all of the projects. The code in shared already depends on
Hammer heavily so it needs to be there. Vis Util can export this
together with the other shared stuff.

* refactor(shared): turn into an encapsulated module

This way it will be quite easy to import from Vis Util later on.